refactor: centralize event mapping
